Utils refactored
