+++ VultisigApp/VultisigApp/iOS/Extension/ViewExtension+iOS.swift
@@ -1,0 +1,28 @@
+//
+//  ViewExtension+iOS.swift
+//  VultisigApp
+//
+//  Created by Amol Kumar on 2024-09-27.
+//
+
+#if os(iOS)
+import SwiftUI
+
+extension View {
+    func navigationBarTitleTextColor(_ color: Color) -> some View {
+        let uiColor = UIColor(color)
+        UINavigationBar.appearance().titleTextAttributes = [.foregroundColor: uiColor ]
+        UINavigationBar.appearance().largeTitleTextAttributes = [.foregroundColor: uiColor ]
+        return self
+    }
+    
+    func detectOrientation(_ orientation: Binding<UIDeviceOrientation>) -> some View {
+        modifier(DetectOrientation(orientation: orientation))
+    }
+    
+    func hideKeyboard() {
+        let resign = #selector(UIResponder.resignFirstResponder)
+        UIApplication.shared.sendAction(resign, to: nil, from: nil, for: nil)
+    }
+}
+#endif
