+++ VultisigApp/VultisigApp/macOS/Utils/OnDropQRUtils.swift
