+++ VultisigApp/VultisigApp/macOS/Utils/Placement+macOS.swift
@@ -1,27 +1,18 @@
 //
-//  Placement.swift
+//  Placement+macOS.swift
 //  VultisigApp
 //
-//  Created by Johnny Luo on 25/6/2024.
+//  Created by Amol Kumar on 2024-09-27.
 //
-
+#if os(macOS)
 import SwiftUI
-enum Placement{
+
+enum Placement {
     case topBarLeading
     case topBarTrailing
     case principal
 
     func getPlacement() -> ToolbarItemPlacement {
-#if os(iOS)
-        switch self {
-        case .topBarLeading:
-            return ToolbarItemPlacement.topBarLeading
-        case .topBarTrailing:
-            return ToolbarItemPlacement.topBarTrailing
-        case .principal:
-            return ToolbarItemPlacement.principal
-        }
-#elseif os(macOS)
         switch self {
         case .topBarLeading:
             return ToolbarItemPlacement.navigation
@@ -30,8 +21,6 @@ enum Placement{
         case .principal:
             return ToolbarItemPlacement.principal
         }
-#endif
     }
 }
-
-
+#endif
